# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPets",
        "description": "bark",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "invited-events/accept": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Accept or decline invite",
                "description": "Accept or decline invitation to event.",
                "operationId": "inviteAccept",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "accept",
                                    "invite_id"
                                ],
                                "properties": {
                                    "accept": {
                                        "description": "Must be 1 or 0. 1 - Accept, 0 - Decline",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "invite_id": {
                                        "description": "Pet invite ID",
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error or registration error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated.|Sign-Up steps not done."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Exist feedback from invite error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invite already accepted.|Invite already declined."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found exseption",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query result from model."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}